Initial commit
This is fedfs-utils-0.6.0, the first public release of fedfs-utils.

Copyright 2010, 2011 Oracle.  All rights reserved.

fedfs-utils is free software; you can redistribute it and/or modify
it under the terms of the GNU General Public License version 2.0 as
published by the Free Software Foundation.

fedfs-utils is distributed in the hope that it will be useful, but
WITHOUT ANY WARRANTY; without even the implied warranty of
MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
GNU General Public License version 2.0 for more details.

You should have received a copy of the GNU General Public License
version 2.0 along with fedfs-utils.  If not, see:

  http://www.gnu.org/licenses/old-licenses/gpl-2.0.txt

Signed-off-by: Chuck Lever <chuck.lever@oracle.com>