mount.fedfs: Overhaul mount.fedfs CLI
The 0.6.0 mount.fedfs interface is not adequate to support mounting
arbitrary points in the global FedFS namespace at arbitrary points
on a client.

To accomodate this functionality, update the mount.fedfs CLI again.

For a synopsis that looks like this:

   mount.fedfs remotedir localdir

"remotedir" will now be a globally useful name (the full path that
starts with "/nfs4/domainname") that indicates what part of the
FedFS namespace to mount, and "localdir" will be the mounted-on
dir on the client to mount this on.

Signed-off-by: Chuck Lever <chuck.lever@oracle.com>