libnsdb: Check all NCEs an NSDB might have
When checking for an FSN that doesn't exist, the NSDB library returns
FEDFS_ERR_NONCE, which is incorrect.  It should return
FEDFS_ERR_NOFSN.

The loop that walks the naming contexts looking for NCE prefixes is
not correct.  A separate loop is needed to construct a list of NSDB
container entries first.  Only if this loop fails to discover any NCEs
should we return FEDFS_ERR_NONCE.

This logic now matches the algorithm described in sections 5.2.1 and
5.2.2 of the NSDB protocol draft.

Signed-off-by: Chuck Lever <chuck.lever@oracle.com>