fedfsc: Usage fix: The NSDB name and port is optional on these commands
Specifying the NSDB name and port is optional in the fedfs clients
because these commands can take the default value of these options
from environment variables.  And, even if the environment variable
for the port number doesn't exist, the default is still port 389.

Correct the usage message in a few of these commands to reflect
this.

Signed-off-by: Chuck Lever <chuck.lever@oracle.com>