nsdb-nces: Minor clean ups
The usage message lists a "-e" option that does not appear to be
supported.

Also, make the output formatting a little neater.  Report the name of
the NSDB on the output in case no hostname was specified on the
command line.

Finally, nsdb-nces was always completing with EXIT_FAILURE.  When an
NCE is detected, complete with EXIT_SUCCESS.

Signed-off-by: Chuck Lever <chuck.lever@oracle.com>