nsdb-create-fsl: minor fixes
The "-t" option was recently changed to "-o".  "-p" is no longer
optional, now that the library creates only fedfsNfsFsl records.

The output of nsdb-create-fsl, on success, should report the full
DN of the new record, just as nsdb-create-fsn now does.

At some later point we need to fix the libnsdb API to allow callers
to specify the subtype of FSL they want to create.

Signed-off-by: Chuck Lever <chuck.lever@oracle.com>