Clean up cert file creation
Narrow permissions on created cert files.  I think only the "fedfs"
user ID and root should be able to read or write these.

While we're here, use symbolic constants to specify the permission
setting.

NB: This code does not work before or after this patch is applied, as
it is currently provisional and untested.  More is planned for a
subsequent release.

Signed-off-by: Chuck Lever <chuck.lever@oracle.com>