Explain how to create a domain root on a Linux NFS server
Update fedfs(7) and INSTALL to describe how to create a domainroot
directory on a Linux NFS server.  This is somewhat provisional, as
we are planning down the road to replace the /etc/exports referral
mechanism.  But it really should be explained somewhere in the
meantime.

Instructions provided by a sleepless Bruce over a beer.

Signed-off-by: Chuck Lever <chuck.lever@oracle.com>