nsdbparams: Specify fedfs UID and GID by name
The --uid and --gid command line options for nsdbparams both take a
number.  For flexibility, also allow text user and group IDs.

This logic is copied from rpc.fedfsd.  It provides better checking
in the case when a UID or GID is specified that is not known to the
system.

Signed-off-by: Chuck Lever <chuck.lever@oracle.com>