libnsdb: Add an API to ping an NSDB
Suppose a program wants to see if an LDAP server is an NSDB.  It
should be up and it should respond to LDAP requests.  We should
be able to get list of namingContexts, and at least one of the
contexts should have an NCE prefix.

This is a procedure that may be done before an NSDB is registered
in the local NSDB connection parameter database.  We introduce two
APIs: one that takes a hostname and port, and another that takes
an instatiated nsdb_t.

Signed-off-by: Chuck Lever <chuck.lever@oracle.com>