nsdbparams: Warn when adding an NSDB to local db that isn't an NSDB
Display a warning when adding NSDB connection parameters for an LDAP
server that isn't already an NSDB.  This is a courtesy warning.

We have a chicken and egg problem.  nsdb-create-nces won't work if
the underlying NSDB connection parameters database doesn't contain
the NSDB in question.  So nsdbparams has to let in entries which
aren't yet NSDBs.

The NSDB draft requires, however, that our rpc.fedfsd prevent setting
NSDB connection parameters for NSDBs that are not NSDBs yet.

Signed-off-by: Chuck Lever <chuck.lever@oracle.com>