fedfsc: Add man pages for the FedFS ADMIN clients
The FedFS ADMIN client commands are meant to be "micro" commands for
testing the FedFS ADMIN protocol and our our implementation of it.
They can also be used in scripts.  I expect this functionality to be
subsumed eventually into richer high level administrative commands.

To enable their wider use in the meanwhile, document them in man
pages.

Signed-off-by: Chuck Lever <chuck.lever@oracle.com>