libnsdb: Return FEDFS_ERR_INVAL for bad parameters
When I originally coded this, I used FEDFS_ERR_SVRFAULT when an API
function encounted a bad parameter.  This is because a bad parameter
would definitely be a software bug in the fedfs-utils server code.

However, in the longer run, we're considering making these functions
available to any application.  So the API should return a more
conventional error code when a function is passed an unusable
paramater.

We keep FEDFS_ERR_SVRFAULT for memory allocation failures.

In addition, all the public API functions should verify that the
ldap_err pointer is not NULL, and return FEDFS_ERR_INVAL if it is.

Signed-off-by: Chuck Lever <chuck.lever@oracle.com>