nsdb-update-nci: NSDB admin client to update NSDB container info
Currently the nsdb-create-nce command and the underlying libnsdb API
don't correctly handle the case where the NCE prefix is not empty.
The NSDB container info is set correctly, but the actual NSDB
container entry is not created.

I'd like this to be a simple, flexible way to turn an LDAP server into
an NSDB.  It doesn't have to cover every case.  Some corner cases
would be hard; for example, what if the NCE's objectClass attribute is
incorrect?  What if there are missing intermediate non-leaf entries
between the namingContext and the NCE?

So we may be better off simply managing the NSDB container information
and leaving the creation of these objects (which are more or less
shared with the rest of the LDAP server) to the LDAP administrator.
This allows great flexibility: the administrator can set these objects
up any way she likes, and then simply drop the NCE prefix into the
naming context.

As nsdb-update-nci replaces nsdb-create-nce and nsdb-delete-nce,
these commands (and the underlying APIs) are removed.

Signed-off-by: Chuck Lever <chuck.lever@oracle.com>