nsdbc: Add man pages for the NSDB clients
The NSDB client commands are meant to be "micro" commands for testing
the NSDB draft protocol and our our implementation of it.  They can
also be used in scripts.  I expect that this functionality to be
subsumed eventually into richer high level administrative commands.

To enable their wider use in the meanwhile, document them in man
pages.

Signed-off-by: Chuck Lever <chuck.lever@oracle.com>
[ jlayton: fix doc/man/Makefile.am to include nsdb-* man pages ]
Signed-off-by: Jeff Layton <jlayton@redhat.com>