autotools: install mount.fedfs in /sbin
/bin/mount expects to find mount helpers in /sbin, so set up the code to
install it there. We can't, however redefine the code to install it in
/sbin unconditionally as that will break "make distcheck".

This patch instead borrows the scheme used by util-linux-ng. If the
prefix is '/usr' or NONE, then we set root_sbindir to '/sbin'. If it's
anything else, then it is set to the same value as the sbindir.

That allows "make distcheck" to pass and still allows mount.fedfs to
be installed in the right place for distro packaging tools.

Signed-off-by: Jeff Layton <jlayton@redhat.com>