libjunction: Clear mode bits when creating a junction
Because NFSD has to examine each directory in a path when doing
pathname resolution, the test to see if a directory is a junction must
be very efficient.  Retrieving an inode's extended attributes is not
fast, so we use a special combination of mode bits to signal that it's
going to be worth the expense of looking at the xattrs during pathname
resolution.

Up to this point, libjunction was simply setting the sticky bit on the
directory.  The sticky bit by itself is fairly common amongst
directories, however.  The original design was to clear the permission
bits as well, since zero permission with the sticky bit set is
nonsense, and also prevents local readers from seeing what's inside
the junction directory.

However, Trond's NFSD patch wasn't looking at the permission bits at
all.  So my libjunction implementation never bothered to clear a new
junction's permission bits.  Now that we are trying to merge the NFSD
patches upstream, we've discovered the omission.  The merged version
of these patches will consider the permission bits, therefore
libjunction must clear them when setting up a junction.

Signed-off-by: Chuck Lever <chuck.lever@oracle.com>