configure.ac: Add a --with-fedfsuser and --with-statedir options
Allow distributions to set the name of the user ID and group ID used
for FedFS system operations and to set the directory used to contain
TLS certs and the NSDB parameter database.

An install hook to update man pages with the directory pathname is
also introduced.  I noticed that subdirectory pathname for storing
certs is incorrect in the man pages; fixed.

Signed-off-by: Chuck Lever <chuck.lever@oracle.com>