nsdbparams: Delete subcommand should report missing NSDB entry
Currently, the "delete" subcommand reports that an NSDB entry
in the connection parameter database was successfully removed
even if the requested NSDB entry was not there to begin with.
This is somewhat confusing.

Have "nsdbparams delete" report when the NSDB entry is missing.
This is still treated as a success.

Signed-off-by: Chuck Lever <chuck.lever@oracle.com>