fedfsc: utf8string's are not NUL-terminated
Take care when printing utf8strings.  They are not NUL-terminated.
Even if there is a NUL character after the string, it was not
allocated by malloc(3), so allowing printf("%s") to walk down these
strings until it finds a NUL will cause it to run into bad memory.

The solution to this is to ensure a printf precision formatter is
always used when displaying utf8string data.