nsdbc: nsdb-list displays "(null)" for the NCE
If no NCE is specified on the command line, the user wants the command
to hunt for any NCE on the NSDB.  However, nsdb-list displays this:

  $ ./nsdb-list
  NSDB: nsdb.nfs4bat.org:389, (null)
    FSN UUID: d3af69ba-f90e-11e0-9390-000c2979d1f3
      FSL UUID: 4c054068-f911-11e0-ac05-000c2979d1f3

The "(null)" is unhelpful.

In the future, it would be nice to display the NCE under which FSNs
were found.  That might be quite a bit of work, as it would require
bringing the loop that walks through the naming contexts out of the
library and adding it to the nsdb-list command.

So I'm going to put that off for now.

Signed-off-by: Chuck Lever <chuck.lever@oracle.com>