libnsdb: nsdb_list_s() returns FEDFS_OK without updating "fsns"
Trying the nsdb-list command against an empty NSDB causes a
segmentation fault.  This is because nsdb_list_s() found no FSN
records.  It did not alter the "fsns" output argument, but returned
FEDFS_OK anyway.

nsdb_list_s() needs to return an appropriate error code if it does not
plant anything in the "fsns" output argument.  To address this, copy
the logic from the tail of nsdb_resolve_fsn_s().

And, fix the nsdb-list command to recognize FEDFS_ERR_NSDB_NOFSN and
report an appropriate error message.

Signed-off-by: Chuck Lever <chuck.lever@oracle.com>