libnsdb: Fix memory allocation and access bugs in nsdb_normalize_path()
The malloc(3) call in nsdb_normalize_path() neglected to allocate
space for the NUL on the end of the string.

As a minor optimization, we don't need to perform the strlen(3) at the
end of the function to compute the length of the result.  We already
have that value in j.

Signed-off-by: Chuck Lever <chuck.lever@oracle.com>