libnsdb: Eliminate nsdb_create_basedir()
By having nsdbparams retain cap_dac_override=ep, we can create the
state directory after dropping privileges, rather than before.  This
gives better error reporting when fedfsd or nsdbparams are not run as
root, and it also automatically creates the directory with the correct
owner and group (the extra passwd look-up and chown(2) is
unnecessary).

A nice side effect is that the state directory can now be created as
needed right in nsdb_init_database() rather than having a separate
public function to do it.  Thus nsdb_create_basedir() is no longer
needed and can be removed.

Note that the new behavior is slightly different than the old:
nsdb_create_basedir() used to use FEDFS_USER always.  Now the
directory will get whatever uid and gid might be specified on the
fedfsd or nsdbparams command line.  This makes the --uid and --gid
options behave closer to user expectations.

Signed-off-by: Chuck Lever <chuck.lever@oracle.com>