libnsdb: struct fedfs_fsl object management
Expose a richer API for managing struct fedfs_fsl objects.  Today
these are returned by the FSN resolution API, thus they are created
only inside the library, on behalf of nsdb_resolve_fsn_s().

But we also want to use them to allow callers outside of libnsdb to
pass in a full set of locations information when creating FSL records.

Signed-off-by: Chuck Lever <chuck.lever@oracle.com>