libnsdb: Update nsdb_create_fsl_s() API
nsdb_create_fsl_s() needs some serious updating.

First, it doesn't make any sense from FedFS's point of view to create
a fedfsFsl record which is not for a specific underlying file system
type.  Remove the ability to create a "fedfsFsl only" record.

Next, the NFS FSL records it does create now contain just default
values.  It was expected that then an administrator would come back to
the new FSL record with nsdb-update-fsl to adjust the settings as
needed.

But this is awkward when setting up lots of records with
fs_locations_info data.  So, instead of creating an NFS FSL with
default values, let's have this API take a fully initialized struct
fedfs_fsl.  The caller can then specify all of the settings at the
time the record is created.

Finally, struct fedfs_fsl objects can be part of a list.  Thanks to
the previous change, we can now pass a list of these objects to
nsdb_create_fsl_s() in order to set up multiple FSL records in a
single call.  Or a list containing a single fedfs_fsl can be passed
in so it will work just as we do today.

Because of this final change, rename the function nsdb_create_fsls_s().

For now, nsdb_create_fsls_s() will not handle setting up the
fedfsAnnotations and fedfsDescr attributes.  That will need a little
more research.  The original API didn't change these attributes
either, so this is not a regression.

Signed-off-by: Chuck Lever <chuck.lever@oracle.com>