libpath: Merge libpath.a into libnsdb.a
The original idea of having a separate libnsdb.a and libpath.a was
to not link libpath.a where it wasn't being used.

However, all of the path functions are called "nsdb_something" and the
whole point of the path XDR functions is to store pathnames in FSL
records, which is entirely an NSDB construct.

What's more, keeping these separate has meant no end of linking
headaches over the months due to cross dependencies between libpath
and libnsdb.  So these two things go together: It's pretty hard to use
libpath.a without libnsdb.a

Let's make things simpler and just merge libpath.a back into
libnsdb.a.  It's just a single source file.

Signed-off-by: Chuck Lever <chuck.lever@oracle.com>