libnsdb: Introduce new path_array utility functions
Since the beginning, I've tried to stick with representing local
pathnames using a C string, and converting to the array form only when
I need to transmit the pathname off-system.  That seemed more natural
for operating on Linux, whose VFS is POSIX-style.

However, if we want to allow creation of junctions that point to
servers that don't necessarily use '/' as the pathname component
separator, we will need a more general way to represent pathnames in
struct fedfs_fsl and in other places.

This commit adds pathname parsing functions in src/libnsdb/path.c
that can handle a new pathname type, which is an array of character
pointers.  Each pointer is to a single component of the pathname.

This data structure parallels FedFsPathName and the equivalent
data structures in NFS's fs_locations4 and fs_locations_info4.

Aside from exposing the richer fs_locations data structure to FedFS
callers, this change moves an important part of pathname resolution
policy out of libnsdb and into callers.  It should also provide an
opportunity for stronger and more consistent checks on incoming
pathnames.

During testing it was discovered that the old pathname parsing logic
is full of bugs, and that these new routines may not be completely
compatible.  This may mean that reading some existing FSLs may result
in errors, and therefore they must be replaced.  However, the new
routines are more likely to interoperate with other FedFS
implementations.

Signed-off-by: Chuck Lever <chuck.lever@oracle.com>