fedfsc: Use new path_array APIs in the FedFS ADMIN clients
Since the beginning, I've tried to stick with representing local
pathnames using a C string, and converting to the array form only when
I need to transmit the pathname off-system.  That seemed more natural
for operating on Linux, whose VFS is POSIX-style.

I've introduced a new pathname API based on arrays of components.
Convert the FedFS ADMIN clients to use this new API when dealing with
input pathnames.

This change moves an important part of pathname resolution policy out
of libnsdb and into callers.  It should also provide an opportunity
for stronger and more consistent checks on incoming pathnames.

There shouldn't be any change in behavior with this patch.  Someday,
though, we may want to expand the command line interface of these
clients (or, build new clients) that allow more precise specification
of the pathname arguments.

Signed-off-by: Chuck Lever <chuck.lever@oracle.com>