libnsdb: Use path array for fedfsNfsPath attribute in nsdb_update_fsl_s()
Since the beginning, I've tried to stick with representing local
pathnames using a C string, and converting to the array form only when
I need to transmit the pathname off-system.  That seemed more natural
for operating on Linux, whose VFS is POSIX-style.

However, if we want to allow creation of junctions that point to
servers that don't necessarily use '/' as the pathname component
separator, we will need a more general way to represent pathnames in
struct fedfs_fsl and friends.

That general way to represent pathnames is as an array of component
character strings.  This is, after all, what is done for fs_locations
in NFS.

Change the nsdb_update_fsl_s() API to expect an array of components
when asked to change the fedfsNfsPath attribute of an NFS FSL record.

Aside from exposing the richer fs_locations data structure to FedFS
callers, this change moves an important part of pathname resolution
policy out of libnsdb and into callers.  It should also provide an
opportunity for stronger and more consistent checks on incoming
pathnames.

Signed-off-by: Chuck Lever <chuck.lever@oracle.com>