Use arrays to represent pathnames in struct fedfs_fsl
Since the beginning, I've tried to stick with representing local
pathnames using a C string, and converting to the array form only when
I need to transmit the pathname off-system.  That seemed more natural
for operating on Linux, whose VFS is POSIX-style.

However, if we want to allow creation of junctions that point to
servers that don't necessarily use '/' as the pathname component
separator, we will need a more general way to represent pathnames in
struct fedfs_fsl.

That general way to represent pathnames is as an array of component
character strings.  This is, after all, what is done for fs_locations
in NFS, and is how pathnames are stored and passed in FedFS.

Thus, for example, to allow callers full expression of the FSL data
type, a plug-in API would express pathnames using an array of
components instead of a simple string containing slashes.

Aside from exposing the richer fs_locations data structure to FedFS
callers, this change moves an important part of pathname resolution
policy out of libnsdb and into callers.  It should also provide an
opportunity for stronger and more consistent checks on incoming
pathnames.

Signed-off-by: Chuck Lever <chuck.lever@oracle.com>