Makefile: Enable building shared libraries
The new junction plug-ins will be shared libraries.  You can't build
a shared library from static libraries, so all of the libraries we
have in fedfs-utils need to be converted to shared libraries.

The use of "-fPIC" as a flag on $(LD) is incorrect, so I moved it to
$(CFLAGS) where needed.  "-fPIE" is appropriate for finished
executables.

As a final touch I converted all of the existing commands to grab the
shared version instead of the static version of each library.

Signed-off-by: Chuck Lever <chuck.lever@oracle.com>