libjunction: Add support for nfs-basic junctions
Now that we have some example XML support in libjunction, we can begin
to sketch in support for NFS junctions.  Instead of storing a FedFS
fileset name which is resolved to a set of fileset locations, an NFS
junction stores the actual set of fileset locations.

The XML for an NFS junction looks something like this:

  <?xml version="1.0" encoding="UTF-8"?>
  <junction>
    <fileset>

      <location>
        <host>fileserver.example.net</host>
        <path>
          <component>foo</component>
          <component>bar</component>
          <component>baz</component>
        </path>
        <version major="4" minor="0" />
        <currency>-1</currency>
        <genflags writable="false" going="false" split="true" />
        <transflags rdma="true" />
        <class simul="0"
               handle="0"
               fileid="0"
               writever="0"
               change="0"
               readdir="0" />
        <read rank="0" order="0" />
        <write rank="0" order="0">
        <flags varsub="false" />
        <validfor>0</validfor>
      </location>

      ....

    </fileset>
  </junction>

Basically the <location> element is a FedFS NFS fileset location
record, expressed in XML.  A list of these is parsed out of a junction
XML document and then presented to, say, mountd.

Although NFS supports both fs_locations and fs_locations_info, there
is only one type of NFS fileset location record.  This is exactly the
same kind of information mountd would get if it resolved a FedFS FSN.
Non-3530 information can be discarded to construct a regular
fs_locations in response to an NFSv4.0 client.

Signed-off-by: Chuck Lever <chuck.lever@oracle.com>