plugin: Introduce the junction plug-in API
Define an external API/ABI to allow applications to access NFS
location data stored in junctions.  This interface is specific to NFS
to continue to provide strong data and function type checking.  Other
file system protocols will need to define their own API.

A junction plug-in implementation is also provided.

Note that the interface currently returns a flat POSIX pathname rather
than an array of components.  The only consumer of this interface so
far is mountd, who is not currently capable of passing an array of
components to the kernel.  To keep things simple for now, I'm going to
handle pathname mangling inside the plug-in rather than in mountd.

At some later point, a second, richer, "get location" method can be
added that gives access to all of the fields in each location record,
to support fs_locations_info4.  At that time support for pathname
component arrays can be introduced.

Signed-off-by: Chuck Lever <chuck.lever@oracle.com>