nsdbc: Exit with FEDFS_ERR_INVAL if default settings don't exist
Commit afc64d55 changed the NSDB client commands to exit with a FedFS
status code.

One case that was missed was when the commands can't find the value of
the LDAP bind DN or NCE from the command line, from environment
variables, nor from the local NSDB connection parameter database.  In
this case, the commands were exiting with FEDFS_OK.

Change these commands to instead exit with FEDFS_ERR_INVAL (an invalid
parameter was provided) in this case.

Signed-off-by: Chuck Lever <chuck.lever@oracle.com>