nfsref: "nfsref remove" should remove FSN found in junction
I have been told that, when removing a FedFS junction, the Solaris
version of the nfsref command deletes the FSN it finds in a
junction.

It is assumed that in general, each junction created by the nfsref
command will refer to its own FSN (ie, FSNs will not be shared by
multiple nfsref- created junctions).

Change the "nfsref remove" subcommand to work the way nfsref
currently works on Solaris.

Signed-off-by: Chuck Lever <chuck.lever@oracle.com>