libjunction: Save directory mode in junction XML
Today, when we convert a directory into a junction, we save the
directory's mode bits in an extended attribute, and then set them to
a known weird value, which the kernel can detect.

As an experiment, save the directory mode in the junction XML as
well.  If this works out, we can avoid using a separate extended
attribute for the saved mode.

Signed-off-by: Chuck Lever <chuck.lever@oracle.com>