nsdbc: Remove default NCE of "o=fedfs"
The man pages for the nsdb-update-nci and nsdb-remove-nci commands
do not mention a default NCE (in fact, they say that the commands
fail if no NCE is specified).

The NSDB protocol specification hints that "o=fedfs" is the preferred
default, but otherwise there's no good reason to choose that as a
default value.  Remove it.