libnsdb: Add nsdb_find_naming_context_s()
Thanks to recent changes in the NSDB protocol, we don't need to
compute a DN prefix any more, so introduce a simpler API.

I've prefer to have the LDAP server tell us which root suffix is the
parent of a DN, but I haven't found a way to do this.  The correct
root suffix is still chosen by iteratively comparing the DN's suffix
with the root suffix.

Signed-off-by: Chuck Lever <chuck.lever@oracle.com>