libnsdb: Generalize nsdb_parse_annotations()
At one point, nsdb_parse_annotations() used to parse fedfsAnnotations
by breaking them into key-value pairs.  But now it just stores each
pair into an array of strings.  This was changed before the initial
0.6.0 commit, but the documenting comment was never updated.

This looks like something that can be used for any FedFS-related
LDAP object that has a fedfsAnnotations attribute.  Instead of
having a version of this function that is specific to fedfsFsl
records, let's make it more general by having it output a string
array rather than filling in a struct fedfs_fsl.

Signed-off-by: Chuck Lever <chuck.lever@oracle.com>