Replace fedfsFslHost with fedfsNfsURI
The latest NSDB schema moves the name and port of the fileserver
host (fedfsFslHost and fedfsFslPort) to the NFS-specific section of
the FSL, and combines them with the export pathname.  The result is
stored in a single labeledURI type attribute.

We'll need to replace the logic that stores the pathname as an XDR
blob on the NSDB with one that constructs an NFS URI.  The logic
that reads an XDR-encoded pathname and unmarshals it will have to
read the URI and break it into its constituent components.

Note: this commit adds a new library dependency to fedfs-utils:  URI
parsing is done via the open liburiparser library.

Signed-off-by: Chuck Lever <chuck.lever@oracle.com>