nsdbc: Introduce a tool for creating a simple NSDB.
NCEs like "o=fedfs", where the NCE and the naming context are the
same entry, are allowed and supported.  But we need something that
makes it easy for a novice to get started.

Add a new tool called nsdb-simple-nce that creates an entry as a
child of some parent DN (usually the LDAP server's domain naming
context, ie. dc=example,dc=net), and then makes that new entry an
NCE for its naming context.

Signed-off-by: Chuck Lever <chuck.lever@oracle.com>