contrib: Add tools to help configure LDAP servers as NSDBs
Setting up an NSDB takes a bunch of detailed steps.  To encourage
adoption of FedFS (or at least broaden our base of testers), we want
to make NSDB set up easier to do.  One of the challenges is that
each LDAP server implementation requires different administrative
tools and procedures.

Thus, for each LDAP server implementation, we can add tools that
set up the FedFS schema, create root suffixes, and so on.  Of course
some server implementations will not succumb to this approach, as
they will still require manual set up on the servers themselves.  We
can but try.

As a first step, create a landing place for server-specific tools
and introduce a few sample tools to remotely configure LDAP servers.

Signed-off-by: Chuck Lever <chuck.lever@oracle.com>