libnsdb: Proper rationalization for disabling LDAP_OPT_REFERRALS
It wasn't always the case, but these days, the FedFS drafts _do_
have something to say about how to handle LDAP referrals when
performing an NSDB request.

NSDB clients must authenticate every NSDB they encounter.  Thus an
LDAP referral means the client has to consult its NSDB connection
parameter database again for the referred-to NSDB node in order to
authenticate it.

The LDAP_OPT_REFERRALS option enables the LDAP library to handle
LDAP referrals transparently to the caller.  It's set to LDAP_OPT_ON
by default.  This means libnsdb would never see an LDAP referral.

Since we need to make sure proper authentication occurs in this
case, libnsdb must see LDAP referrals and deal with them.  Thus, the
LDAP_OPT_OFF setting is correct, but the reason in the comment for
that setting is invalid.

Our implementation also has a per-NSDB setting that allows an
administrator to disable following LDAP referrals entirely.  That's
implemented by checking that setting when a referral occurs.  This
logic also requires that the LDAP library expose referrals to
libnsdb.

Signed-off-by: Chuck Lever <chuck.lever@oracle.com>