libnsdb: Add API for retrieving certificate data
We're about to optimize nsdb_lookup_nsdb() for the common case,
which is the case that does not need to retrieve any security data.
In fact, right now, there is only a single caller of
nsdb_lookup_nsdb() that needs this data.

Introduce a separate API for retrieving certificate data for an
already-initialized nsdb_t object.

Signed-off-by: Chuck Lever <chuck.lever@oracle.com>