libnsdb: New mechanism for setting NSDB connection security
The nsdbparams "update" subcommand shouldn't wipe security data when
it is invoked.  Every time an NSDB connection parameters update is
done to change other connection parameters, the previous security
settings for that NSDB are wiped.

It's always worked this way, but until now, fedfs-utils has
supported only one connection security type, a type which had no
security data associated with it.

Also, the LDAP library wants certificate files in PEM format, but
the ADMIN protocol states that certificate material transmitted on
the network (via FEDFS_SET_NSDB_PARAMS, for instance) must be DER
encoded.

In addition to transforming certificate material between PEM and
DER encoding, fedfsd and the ADMIN client tools ought to be doing
some level of validation of user-provided certificate material
before storage and use.

To address these issues, introduce a new API for changing NSDB
connection security settings.  The old API used the
nsdb_update_nsdb() function to both create new entries in the local
NSDB connection parameters database, and to change/update their
security settings.  The new API breaks nsdb_update_nsdb() into two
phases:

  o  The new nsdb_create_nsdb() function creates a new entry in the
     local NSDB connection parameters database

  o  The new nsdb_connsec_set_foo() functions replace the
     security settings in an entry

The purpose of the new arrangement is to split out the details of
certificate management from the details of creating an entry in the
database.  Setting the connection security mode is the same class of
operation as updating any of the other per-NSDB settings, with the
additional complexity of having to deal with the private certificate
files.

To manage certificate encoding transformations, new APIs are
introduced that allow callers to pass just the pathname of the
certificate file to libnsdb, which handles the details of validating
and storing the certificate material.

In subsequent commits, I'll update the various tools and daemons to
use the new API, then finally remove the old API.

Signed-off-by: Chuck Lever <chuck.lever@oracle.com>