libnsdb: nsdb_ping_s() should handle TLS-only NSDBs
If an NSDB requires TLS (as recommended by the NSDB protocol draft)
nsdb_ping_s() will return FEDFS_ERR_NSDB_LDAP_VAL.  When performing
an NSDB ping, an NSDB client doesn't yet have NSDB connection
parameters, so it can't actually connect to the NSDB and see if it
has an NCE.

But nsdb_ping_s() can report in this case that TLS is required.

The two cases where nsdb_ping_s() is invoked are when fedfsd or
nsdbparams is trying to determine if a new NSDB entry is allowed.
If the ping tells us we need TLS security, fedfsd and nsdbparams can
ensure that the new connection parameters do configure FEDFS_SEC_TLS.

Signed-off-by: Chuck Lever <chuck.lever@oracle.com>