nfsref: FSNs not removed if "nfsref" junction type not specified
Suppose "foobar" is an nfs-fedfs junction.

If I explicitly specify the junction type on the command line with
"nfsref remove -t nfs-fedfs foobar", the FSN listed in the junction
"foobar" is deleted on the NSDB.  If I let "nfsref" guess what kind
of junction "foobar is with "nfsref remove foobar", the FSN is not
removed.

Remove the junction's FSN when the junction type is not specified.

nfsref_remove_unspecified() is moved to avoid adding a forward
declaration for nfsref_remove_delete_fsn().

Signed-off-by: Chuck Lever <chuck.lever@oracle.com>