l10n: Replace nl_langinfo(CODESET) check in command line tools
SteveD reports:
> The nfsref *always* fails with "Failed to set locale and langinfo"
>
> Steps to Reproduce:
> 1. type nfsref
>
> Additional info:
> The failure is be caused by the following checks:
>
>     /* Ensure UTF-8 strings can be handled transparently */
>     if (setlocale(LC_CTYPE, "") == NULL ||
>         strcmp(nl_langinfo(CODESET), "UTF-8") != 0) {
>         fprintf(stderr, "Failed to set locale and langinfo\n");
>         goto out;
>     }
>
> In f19
>   setlocale(LC_CTYPE, "") == 'en_US
> and
>   nl_langinfo(CODESET) == ISO-8859-1

The failing check was introduced in commit 0520ee72 "Initial
commit", Tue Mar 29 15:37:40 2011.  The point of that check is to
force the use of the UTF-8 codeset, since that codeset is required
by the FedFS protocol specification, but it appears to be based on
an incorrect assumption.

This assumption has worked silently until now.  Apparently Fedora
19 has changed the default shell LANG setting from "en_US.utf8" to
"en_US.iso88591" .

For now we are better off just taking the user's locale and then
determining over time where our libraries need to convert user data
to and from UTF-8 prior to network communication.

Replace the above logic everywhere in fedfs-utils with this:

	(void)setlocale(LC_ALL, "");

BugLink: https://bugzilla.redhat.com/show_bug.cgi?id=969392
Reported-by: Steve Dickson <steved@redhat.com>
Signed-off-by: Chuck Lever <chuck.lever@oracle.com>