Makefile: Specify as-needed when building executables
The "--as-needed" linker flag reduces the number of unused objects
and dependencies in an executable, making it smaller and faster to
load.

Some distributions implicitly specify "--as-needed" when linking
their package builds.  Building on Fedora appears not to.

So that building on any platform can catch build problems before
they are committed, explicitly specify --as-needed for executables.

References:

  http://www.gentoo.org/proj/en/qa/asneeded.xml#doc_chap1

  https://sigquit.wordpress.com/tag/gnu-libtool/

Reviewed-by: David Disseldorp <ddiss@suse.de>
Signed-off-by: Chuck Lever <chuck.lever@oracle.com>