build: fix gcc linker library argument placement
The build currently fails on gcc 4.7.3 based openSUSE systems due to the
placement of library arguments while linking, and the use of the
--as-needed binutils feature.

As described by http://www.gentoo.org/proj/en/qa/asneeded.xml:
For symbols missing in a given file (either an object file, a static
archive or a library), the linker checks only in the files coming after
it.
Without --as-needed, files are linked together without considering
ordering. But with the flag, the libraries that aren't used for
resolving symbols are discarded and thus not linked.

Signed-off-by: David Disseldorp <ddiss@suse.de>